+++ src/common/ExportSettingsModal/ExportSettingsModal.js
@@ -1,10 +1,11 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import {
   Button,
   Label,
+  Loading,
   Modal,
   ModalFooter,
   RadioButton,
@@ -12,12 +13,22 @@ import {
 
 const ExportSettingsModal = ({
   onCancel,
+  isExporting,
+  onExportCSV,
 }) => {
+  const onExport = useCallback(
+    async () => {
+      await onExportCSV();
+      onCancel();
+    },
+    [onExportCSV, onCancel],
+  );
   const exportModalFooter = (
     <ModalFooter>
       <Button
         buttonStyle="primary"
-        disabled
+        onClick={onExport}
+        disabled={isExporting}
       >
         <FormattedMessage id="ui-orders.exportSettings.export" />
       </Button>
@@ -33,29 +44,38 @@ const ExportSettingsModal = ({
       label={<FormattedMessage id="ui-orders.exportSettings.label" />}
       footer={exportModalFooter}
     >
+
       <p><FormattedMessage id="ui-orders.exportSettings.message" /></p>
 
-      <Label>
-        <FormattedMessage id="ui-orders.exportSettings.orderFieldsLabel" />
-      </Label>
-      <RadioButton
-        checked
-        label={<FormattedMessage id="ui-orders.exportSettings.all" />}
-      />
-
-      <Label>
-        <FormattedMessage id="ui-orders.exportSettings.lineFieldsLabel" />
-      </Label>
-      <RadioButton
-        checked
-        label={<FormattedMessage id="ui-orders.exportSettings.all" />}
-      />
+      {isExporting
+        ? <Loading size="large" />
+        : (
+          <>
+            <Label>
+              <FormattedMessage id="ui-orders.exportSettings.orderFieldsLabel" />
+            </Label>
+            <RadioButton
+              checked
+              label={<FormattedMessage id="ui-orders.exportSettings.all" />}
+            />
+            <Label>
+              <FormattedMessage id="ui-orders.exportSettings.lineFieldsLabel" />
+            </Label>
+            <RadioButton
+              checked
+              label={<FormattedMessage id="ui-orders.exportSettings.all" />}
+            />
+          </>
+        )
+      }
     </Modal>
   );
 };
 
 ExportSettingsModal.propTypes = {
   onCancel: PropTypes.func.isRequired,
+  isExporting: PropTypes.bool.isRequired,
+  onExportCSV: PropTypes.func.isRequired,
 };
 
 export default ExportSettingsModal;