+++ src/common/utils/index.js
@@ -8,3 +8,4 @@ export * from './getAddresses';
 export * from './getAddressOptions';
 export { default as getCreateInventorySetting } from './getCreateInventorySetting';
 export * from './getUserNameById';
+export * from './fetchExportDataByIds';